fix: code cleanup
