+++ .github/workflows/main.yml
@@ -1,12 +1,12 @@
 name: "Deployment"
-on: 
+on:
   push:
     branches: [ "master" ]
 jobs:
   deploy-to-gh-pages:
     name: Deploy to github pages
     runs-on: ubuntu-latest
-    steps: 
+    steps:
       - name: Checkout
         uses: actions/checkout@v3
       - name: Setting up Nodejs
