+++ package.json
@@ -8,8 +8,8 @@
     "build": "vite build",
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "predeploy":"vite build --base=/Debojit-Portfolio/",
-    "deploy":"gh-pages -d dist"
+    "predeploy": "vite build --base=/Debojit-Portfolio/",
+    "deploy": "gh-pages -d dist"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.4.0",
