+++ src/Components/Approuter/Approuter.jsx
@@ -1,13 +1,15 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import Navlinks from "../Navbar/Navlinks";
+import Navlinks from "../../assets/data/Navlinks";
 
 export default function Approuter() {
   return (
     <Routes>
-      {Object.values(Navlinks).map((link) => (
-        <Route key={link.to} path={link.to} element={link.element} />
-      ))}
+      {
+        Object.values(Navlinks).map((link) => (
+          <Route key={link.to} path={link.to} element={link.element} />
+        ))
+      }
     </Routes>
   );
 };
