--- src/Components/Contact/Contact.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-
-export default function Contact(){
-    return(
-        <div className="Contact">
-        </div>
-    );
-};
