+++ src/Components/Footer/Footer.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import '../Footer/Footer.css'
+import "./Footer.css";
 
 export default function Footer() {
-    return (
-        <div className="Footer">
-            All Rights Reserved
-        </div>
-    );
+  return (
+    <div className="Footer">
+      All Rights Reserved
+    </div>
+  );
 };
