+++ src/Components/Navbar/Navbar.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import '../Navbar/Navbar.css'
+import "./Navbar.css";
 import { NavLink, Link } from "react-router-dom";
-import Navlinks from "../Navbar/Navlinks";
+import Navlinks from "../../assets/data/Navlinks";
 
-export default function Navbar(){
-    return(
+export default function Navbar() {
+  return (
     <nav className="Navbar">
       <div className="container">
         <div className="logo">
@@ -12,18 +12,18 @@ export default function Navbar(){
             Debojit
           </Link>
         </div>
-      <div className="list">
-        <ul>
-          {
-            Object.values(Navlinks).map((link, index) => (
-              <li key={index}>
-                <NavLink to={link.to}>{link.text}</NavLink>
-              </li>
-            ))
-          }
-        </ul>
-      </div>
+        <div className="list">
+          <ul>
+            {
+              Object.values(Navlinks).map((link, index) => (
+                <li key={index}>
+                  <NavLink to={link.to}>{link.text}</NavLink>
+                </li>
+              ))
+            }
+          </ul>
+        </div>
       </div>
     </nav>
-    );
+  );
 };
