--- src/Components/Portfolio/Portfolio.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-
-export default function Portfolio(){
-    return(
-        <div className="Portfolios">
-        </div>
-    );
-};
