+++ src/pages/About/About.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "./About.css";
+
+export default function About() {
+  return (
+    <div className="About">
+    </div>
+  );
+};
