+++ src/pages/Contact/Contact.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Contact.css";
+
+export default function Contact() {
+  return (
+    <div className="Contact">
+    </div>
+  );
+};
