+++ src/pages/Home/Home.css
@@ -1,11 +1,10 @@
-
 .intro {
   color: white;
   margin-left: 40vh;
   margin-top: 30vh;
   font-size: 3.5vh;
-  font-family: 'Space Mono', monospace;
-  letter-spacing: .15em;
+  font-family: "Space Mono", monospace;
+  letter-spacing: 0.15em;
 }
 
 .icons {
@@ -13,7 +12,7 @@
   font-size: 4vh;
 }
 
-.icons>a {
+.icons > a {
   display: flex;
   color: white;
   border: 2px solid white;
