+++ src/pages/Portfolio/Portfolio.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Portfolio.css";
+
+export default function Portfolio() {
+  return (
+    <div className="Portfolios">
+    </div>
+  );
+};
