+++ src/styles/index.css
@@ -1,5 +1,5 @@
-*{
-  margin:0px;
+* {
+  margin: 0px;
   box-sizing: border-box;
   background-color: black;
 }
